+++ packages/theme/node.d.ts
@@ -1,38 +1,12 @@
-import { Route, DefaultTheme, UserConfig } from 'vitepress';
-import { ElButton } from 'element-plus';
+import { DefaultTheme, UserConfig } from 'vitepress';
 import { RSSOptions } from 'vitepress-plugin-rss';
 import { Repo, Mapping } from '@giscus/vue';
 import { Options } from 'oh-my-live2d';
-import { Ref } from 'vue';
 import { PagefindConfig } from 'vitepress-plugin-pagefind';
+import { AnnouncementOptions } from 'vitepress-plugin-announcement';
 export { tabsMarkdownPlugin } from 'vitepress-plugin-tabs';
 
 type RSSPluginOptions = RSSOptions;
-declare namespace BlogPopover {
-    interface Title {
-        type: 'title';
-        content: string;
-        style?: string;
-    }
-    interface Text {
-        type: 'text';
-        content: string;
-        style?: string;
-    }
-    interface Image {
-        type: 'image';
-        src: string;
-        style?: string;
-    }
-    interface Button {
-        type: 'button';
-        link: string;
-        content: string;
-        style?: string;
-        props?: InstanceType<typeof ElButton>['$props'];
-    }
-    type Value = Title | Text | Image | Button;
-}
 type ThemeableImage = string | {
     src: string;
     alt?: string;
@@ -175,6 +149,16 @@ declare namespace Theme {
          * @default 'sidebar'
          */
         style?: 'card' | 'sidebar';
+        /**
+         * 是否在左侧显示日期
+         * @default true
+         */
+        showDate?: boolean;
+        /**
+         * 是否在左侧展示序号
+         * @default true
+         */
+        showNum?: boolean;
     }
     interface HomeAnalysis {
         articles?: {
@@ -199,6 +183,10 @@ declare namespace Theme {
         analysis?: HomeAnalysis;
     }
     interface ArticleConfig {
+        /**
+         * 文章分析数据展示标题
+         */
+        analyzeTitles?: ArticleAnalyzeTitles;
         readingTime?: boolean;
         /**
          * 阅读时间分析展示位置
@@ -206,6 +194,48 @@ declare namespace Theme {
          */
         readingTimePosition?: 'inline' | 'newLine' | 'top';
         hiddenCover?: boolean;
+    }
+    interface ArticleAnalyzeTitles {
+        /**
+         * 字数：{{value}} 个字
+         */
+        topWordCount?: string;
+        /**
+         * 预计：{{value}} 分钟
+         */
+        topReadTime?: string;
+        /**
+         * {{value}} 个字
+         */
+        inlineWordCount?: string;
+        /**
+         * {{value}} 分钟
+         */
+        inlineReadTime?: string;
+        /**
+         * 文章字数
+         */
+        wordCount?: string;
+        /**
+         * 预计阅读时间
+         */
+        readTime?: string;
+        /**
+         * 本文作者
+         */
+        author?: string;
+        /**
+         * 发布时间
+         */
+        publishDate?: string;
+        /**
+         * 最近修改时间
+         */
+        lastUpdated?: string;
+        /**
+         * 标签
+         */
+        tag?: string;
     }
     interface Alert {
         type: 'success' | 'warning' | 'info' | 'error';
@@ -222,50 +252,6 @@ declare namespace Theme {
         closeText?: string;
         showIcon?: boolean;
         html?: string;
-    }
-    /**
-     * 公告
-     */
-    interface Popover {
-        title: string;
-        /**
-         * 细粒度的时间控制
-         * 默认展示时间，-1 只展示1次，其它数字为每次都展示，一定时间后自动消失，0为不自动消失
-         * 配置改变时，会重新触发展示
-         */
-        duration: number;
-        /**
-         * 移动端自动最小化
-         * @default false
-         */
-        mobileMinify?: boolean;
-        body?: BlogPopover.Value[];
-        footer?: BlogPopover.Value[];
-        /**
-         * 手动重新打开
-         * @default true
-         */
-        reopen?: boolean;
-        /**
-         * 是否打开闪烁提示，通常需要和 reopen 搭配使用
-         * @default true
-         */
-        twinkle?: boolean;
-        /**
-         * 设置展示图标，svg
-         * @recommend https://iconbuddy.app/search?q=fire
-         */
-        icon?: string;
-        /**
-         * 设置关闭图标，svg
-         * @recommend https://iconbuddy.app/search?q=fire
-         */
-        closeIcon?: string;
-        /**
-         * 自定义展示策略
-         * @param to 切换到的目标路由
-         */
-        onRouteChanged?: (to: Route, show: Ref<boolean>) => void;
     }
     interface FriendLink {
         nickname: string;
@@ -336,6 +322,10 @@ declare namespace Theme {
     interface BlogConfig {
         blog?: false;
         /**
+         * 展示日期格式化
+         */
+        formatShowDate?: FormatShowDate;
+        /**
          * 内置一些主题色
          * @default 'vp-default'
          * 也可以自定义颜色，详见 https://theme.sugarat.top/config/style.html#%E4%B8%BB%E9%A2%98%E8%89%B2
@@ -368,7 +358,7 @@ declare namespace Theme {
          * el-alert
          */
         alert?: Alert;
-        popover?: Popover;
+        popover?: AnnouncementOptions;
         friend?: FriendLink[] | FriendConfig;
         authorList?: Omit<FriendLink, 'avatar'>[];
         /**
@@ -441,6 +431,32 @@ declare namespace Theme {
          */
         imageStyle?: ImageStyleConfig;
     }
+    type FormatShowDate = {
+        /**
+         * 刚刚
+         */
+        justNow?: string;
+        /**
+         * 秒前
+         */
+        secondsAgo?: string;
+        /**
+         * 分钟前
+         */
+        minutesAgo?: string;
+        /**
+         * 小时前
+         */
+        hoursAgo?: string;
+        /**
+         * 天前
+         */
+        daysAgo?: string;
+        /**
+         * 周前
+         */
+        weeksAgo?: string;
+    } | ((date: Date | string) => string);
     interface BackToTop {
         /**
          * 距离顶部多少距离出现
@@ -483,7 +499,11 @@ declare namespace Theme {
         /**
          * 是否展示主题版本信息
          */
-        version?: boolean;
+        version?: boolean | {
+            name?: string;
+            link?: string;
+            icon?: boolean | string;
+        };
         /**
          * copyright
          */